+++ CHANGELOG.md
@@ -7,7 +7,7 @@
   - **_Breaking Change_**: remove `Builder.getTokenKinds`.
   - Feat: allow multiple calls for `Builder.state` and `Builder.error`.
   - Feat: add `ITrimmedLexer/IReadonlyTrimmedLexer`.
-  - Feat: `Builder.append/ignore/anonymous` can accept actions with different data bindings in one call.
+  - Feat: `Builder.append` can accept actions with different data bindings in one call.
   - Fix: `Lexer.trimStart` shouldn't collect non-muted errors.
 - Parser
   - **_Breaking Change_**: lexer will be trimmed after every parse. This should improve the performance.